+++ tests/ids-pager/ids-pager.spec.ts
@@ -1,6 +1,6 @@
 import AxeBuilder from '@axe-core/playwright';
 import percySnapshot from '@percy/playwright';
-import { expect } from '@playwright/test';
+import { Locator, expect } from '@playwright/test';
 import { test } from '../base-fixture';
 
 import IdsPager from '../../src/components/ids-pager/ids-pager';
@@ -82,6 +82,77 @@ test.describe('IdsPager tests', () => {
       expect(await page.locator('ids-pager-button ids-icon').nth(1).getAttribute('icon')).toBe('previous-page');
       expect(await page.locator('ids-pager-button ids-icon').nth(2).getAttribute('icon')).toBe('next-page');
       expect(await page.locator('ids-pager-button ids-icon').nth(3).getAttribute('icon')).toBe('last-page');
+    });
+  });
+
+  test.describe('IdsPager functionality tests', () => {
+    let idsPager: Locator;
+    let idsFirst: Locator;
+    let idsPrevious: Locator;
+    let idsInput: Locator;
+    let idsNext: Locator;
+    let idsLast: Locator;
+    let children: Locator[];
+
+    test.beforeEach(async ({ page }) => {
+      idsPager = await page.locator('ids-pager');
+      idsFirst = await idsPager.locator('ids-pager-button[first]');
+      idsPrevious = await idsPager.locator('ids-pager-button[previous]');
+      idsInput = await idsPager.locator('ids-pager-input');
+      idsNext = await idsPager.locator('ids-pager-button[next]');
+      idsLast = await idsPager.locator('ids-pager-button[last]');
+      await idsPager.evaluate((element: IdsPager) => {
+        const dropdown = document.createElement('ids-pager-dropdown');
+        dropdown.setAttribute('page-size', '20');
+        dropdown.setAttribute('slot', 'end');
+        element.appendChild(dropdown);
+      });
+      children = [idsFirst, idsPrevious, idsInput, idsNext, idsLast];
+    });
+
+    test('can set/get disabled status', async () => {
+      expect(await idsPager.evaluate((element: IdsPager) => element.disabled)).toBeFalsy();
+      children.forEach(async (child) => {
+        await expect(child).not.toHaveAttribute('disabled');
+        await expect(child).not.toHaveAttribute('parent-disabled');
+      });
+
+      expect(await idsPager.evaluate((element: IdsPager) => {
+        element.disabled = true;
+        return element.disabled;
+      })).toBeTruthy();
+      children.forEach(async (child) => {
+        await expect(child).toHaveAttribute('disabled');
+        await expect(child).toHaveAttribute('parent-disabled');
+      });
+    });
+
+    test('can set/get pageSize attribute', async () => {
+      const testData = [
+        { data: 5, expected: 5 },
+        { data: '7', expected: 7 },
+        { data: null, expected: 10 } // default page size is 10
+      ];
+
+      expect(await idsPager.evaluate((element: IdsPager) => element.pageSize)).toEqual(20);
+      await expect(idsPager).toHaveAttribute('page-size', '20');
+      children.forEach(async (child) => {
+        await expect(child).not.toHaveAttribute('page-size', '20');
+      });
+
+      for (const data of testData) {
+        expect(await idsPager.evaluate((element: IdsPager, tData) => {
+          element.pageSize = tData.data as any;
+          return element.pageSize;
+        }, data)).toEqual(data.expected);
+        children.forEach(async (child) => {
+          await expect(child).toHaveAttribute('page-size', data.expected.toString());
+        });
+      }
+    });
+
+    test('can set/get pageSizes attribute', async () => {
+
     });
   });
 });